Copy Decl into the DSL module.
