+++ src/GL/Shader/DSL.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE KindSignatures #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RankNTypes #-}
@@ -16,7 +17,11 @@ module GL.Shader.DSL
 , None(..)
 , shaderSources
 , Frag(..)
-, module GL.Shader.Decl
+  -- * Declarations
+, Decl
+, main
+, primitiveIn
+, primitiveOut
 , module GL.Shader.Expr
 , module GL.Shader.Stmt
   -- * Re-exports
@@ -33,6 +38,7 @@ module GL.Shader.DSL
 ) where
 
 import qualified Control.Category as Cat
+import           Control.Monad.Trans.Cont
 import           Data.Coerce
 import           Data.Function (fix)
 import           Data.Functor.C
@@ -41,8 +47,8 @@ import           Data.Kind (Type)
 import           Data.Text.Prettyprint.Doc hiding (dot)
 import           Data.Text.Prettyprint.Doc.Render.String
 import           GHC.Generics hiding ((:.:))
+import qualified GL.Primitive as P
 import qualified GL.Shader as Shader
-import           GL.Shader.Decl hiding (Type)
 import           GL.Shader.Expr
 import           GL.Shader.Stmt
 import           GL.Shader.Vars
@@ -124,3 +130,43 @@ newtype Frag v = Frag { fragColour :: v (Colour Float) }
   deriving (Generic)
 
 instance Vars Frag
+
+
+runDecl :: (a -> Doc ()) -> Decl k a -> Doc ()
+runDecl k = (`runCont` k) . getDecl
+
+newtype Decl (k :: Shader.Stage) a = Decl { getDecl :: Cont (Doc ()) a }
+  deriving (Applicative, Functor, Monad)
+
+raw :: Doc () -> Decl k ()
+raw d = Decl . cont $ \ k -> d <> k ()
+
+main :: Stmt k () -> Decl k ()
+main body = raw (pretty "void" <+> pretty "main" <> parens mempty <+> braces (nest 2 (line <> renderStmt body <> line)))
+
+
+primitiveIn :: P.Type -> Decl 'Shader.Geometry ()
+primitiveIn ty = raw doc where
+  doc = pretty "layout" <+> parens (render ty) <+> pretty "in;" <> hardline
+  render = \case
+    P.Points        -> pretty "points"
+    P.Lines         -> pretty "lines"
+    P.LineStrip     -> pretty "lines"
+    P.LineLoop      -> pretty "lines"
+    P.TriangleStrip -> pretty "triangles"
+    P.Triangles     -> pretty "triangles"
+
+primitiveOut :: P.Type -> Int -> Decl 'Shader.Geometry ()
+primitiveOut ty mx = raw doc where
+  doc = pretty "layout" <+> parens (render ty <> comma <+> pretty "max_vertices" <+> equals <+> pretty mx) <+> pretty "out;" <> hardline
+  render = \case
+    P.Points        -> pretty "points"
+    P.Lines         -> pretty "line_strip"
+    P.LineStrip     -> pretty "line_strip"
+    P.LineLoop      -> pretty "line_strip"
+    P.TriangleStrip -> pretty "triangle_strip"
+    P.Triangles     -> pretty "triangle_strip"
+
+
+renderDecl :: Decl k a -> Doc ()
+renderDecl = runDecl (const mempty)
